Only remove latest.dams.json symlink if it exists (#1985)
Prior versions of the file emitter unconditionally tried to remove the
latest.dams.json symlink. The current version checks the return value
from unlink, which results in a failure if the symlink doesn't exist.

Fix that by only attempting to remove the symlink if the file exists.

+++ userspace/dragent/src/file_emitter.cpp
@@ -147,12 +147,17 @@ bool file_emitter::emit_flush_data_message(const std::shared_ptr<flush_data_mess
 		if(ofile)
 		{
 			const std::string symbolic_link = m_output_dir + "latest.dams.json";
-			if (unlink(symbolic_link.c_str()) != 0)
+			Poco::File sf(symbolic_link);
+
+			if (sf.exists())
 			{
-				g_logger.format(sinsp_logger::SEV_ERROR,
-						"Could not remove symlink to %s (%s)",
-						symbolic_link.c_str(), strerror(errno));
-				return false;
+				if (unlink(symbolic_link.c_str()) != 0)
+				{
+					g_logger.format(sinsp_logger::SEV_ERROR,
+							"Could not remove symlink to %s (%s)",
+							symbolic_link.c_str(), strerror(errno));
+					return false;
+				}
 			}
 
 			if (symlink(filename.c_str(), symbolic_link.c_str()) != 0)
